+++ packages/core/code-loading/webRequire.js
@@ -1,5 +1,8 @@
 const path = require('path')
 
+// use posix versions of path, even in the browser
+const posix = path.posix ? path.posix : path
+
 const getFileExtensionFromString = require('../utils/getFileExtensionFromString')
 
 /* find matching path in inputs
@@ -74,29 +77,29 @@ const makeWebRequire = (filesAndFolders, options) => {
    * The logic is based on the original NODE require() function.
    * @see https://nodejs.org/dist/latest-v12.x/docs/api/modules.html#modules_all_together
    */
-  const _require = (curPath, reqPath) => {
-    // console.log('***** require: cur:', curPath, ' req:', reqPath)
+  const _require = (currentPath, requirePath) => {
+    // console.log('***** require: cur:', currentPath, ' req:', requirePath)
 
     // core modules
-    const directModule = coreModules[reqPath]
+    const directModule = coreModules[requirePath]
     if (directModule) {
       return directModule.exports
     }
 
-    if (!curPath || reqPath.startsWith('/')) {
-      curPath = ''
+    if (!currentPath || requirePath.startsWith('/')) {
+      currentPath = '/'
     }
 
-    const loadAsFile = (reqPath) => {
-      // console.log('***** load as file', reqPath)
-      let baseExt = getFileExtensionFromString(reqPath)
+    const loadAsFile = (requirePath) => {
+      // console.log('***** load as file', requirePath)
+      let baseExt = getFileExtensionFromString(requirePath)
       if (!baseExt) {
         baseExt = 'js' // for lookups
-        reqPath = reqPath + '.js'
+        requirePath = requirePath + '.js'
       }
       baseExt = '.' + baseExt
 
-      const entry = findMatch(reqPath, filesAndFolders)
+      const entry = findMatch(requirePath, filesAndFolders)
       if (!entry) return null
 
       if (entry.children) return null // directory
@@ -118,39 +121,39 @@ const makeWebRequire = (filesAndFolders, options) => {
       return null
     }
 
-    const loadIndex = (reqPath) => {
-      // console.log('***** load index', reqPath)
-      const entry = findMatch(reqPath, filesAndFolders)
+    const loadIndex = (requirePath) => {
+      // console.log('***** load index', requirePath)
+      const entry = findMatch(requirePath, filesAndFolders)
       if (!entry) return null
 
-      if (reqPath === '/') reqPath = '' // FIXME hack for multiple file dragNdrop
+      if (requirePath === '/') requirePath = '' // FIXME hack for multiple file dragNdrop
 
-      let indexPath = reqPath + '/index.js'
+      let indexPath = requirePath + '/index.js'
       let matchingModule = loadAsFile(indexPath)
       if (matchingModule) return matchingModule
 
-      indexPath = reqPath + '/index.json'
+      indexPath = requirePath + '/index.json'
       matchingModule = loadAsFile(indexPath)
       if (matchingModule) return matchingModule
 
       return null
     }
 
-    const loadAsDirectory = (reqPath) => {
-      // console.log('***** load as directory', reqPath)
-      let entry = findMatch(reqPath, filesAndFolders)
+    const loadAsDirectory = (requirePath) => {
+      // console.log('***** load as directory', requirePath)
+      let entry = findMatch(requirePath, filesAndFolders)
       if (!entry) return null
 
       if (!entry.children) return null // file
 
       // load from main definition
       let matchingModule
-      const jsonPath = reqPath + '/package.json'
+      const jsonPath = requirePath + '/package.json'
       entry = findMatch(jsonPath, filesAndFolders)
       if (entry) {
         const main = JSON.parse(entry.source).main
         if (main) {
-          const mainPath = reqPath + '/' + main
+          const mainPath = requirePath + '/' + main
           matchingModule = loadAsFile(mainPath)
           if (matchingModule) return matchingModule
 
@@ -162,24 +165,24 @@ const makeWebRequire = (filesAndFolders, options) => {
       }
 
       // load index
-      matchingModule = loadIndex(reqPath)
+      matchingModule = loadIndex(requirePath)
       if (matchingModule) return matchingModule
 
       return null
     }
 
-    // relative paths
-    if (reqPath.startsWith('./') || reqPath.startsWith('/') || reqPath.startsWith('../')) {
-      reqPath = path.resolve(path.dirname(curPath), reqPath)
+    // relative paths (POSIX style)
+    if (requirePath.startsWith('./') || requirePath.startsWith('/') || requirePath.startsWith('../')) {
+      requirePath = posix.normalize(posix.dirname(currentPath) + posix.sep + requirePath)
       // load as file
-      let loadedModule = loadAsFile(reqPath)
+      let loadedModule = loadAsFile(requirePath)
       if (loadedModule) return loadedModule
 
       // load as directory
-      loadedModule = loadAsDirectory(reqPath)
+      loadedModule = loadAsDirectory(requirePath)
       if (loadedModule) return loadedModule
 
-      throw new Error(`Cannot find relative path to module ${reqPath}`)
+      throw new Error(`Cannot find relative path to module ${requirePath}`)
     }
 
     // TODO load self-reference
@@ -189,18 +192,18 @@ const makeWebRequire = (filesAndFolders, options) => {
       const dirs = []
       for (let i = parts.length - 1; i > 0; i--) {
         if (parts[i] === 'node_modules') continue
-        const dir = path.sep + path.join(...parts.slice(1, i + 1), 'node_modules')
+        const dir = posix.sep + posix.join(...parts.slice(1, i + 1), 'node_modules')
         dirs.push(dir)
       }
       return dirs
     }
 
-    const loadNodeModules = (reqPath, basePath) => {
-      // console.log('loadNodeModules',reqPath, basePath)
+    const loadNodeModules = (requirePath, basePath) => {
+      // console.log('loadNodeModules',requirePath, basePath)
       const dirs = nodeModulesPaths(basePath)
       for (let i = 0; i < dirs.length; i++) {
         const dir = dirs[i]
-        const relPath = path.join(dir, reqPath)
+        const relPath = posix.join(dir, requirePath)
         // load as file
         let loadedModule = loadAsFile(relPath)
         if (loadedModule) return loadedModule
@@ -212,14 +215,14 @@ const makeWebRequire = (filesAndFolders, options) => {
     }
 
     // load node_module
-    const loadedModule = loadNodeModules(reqPath, path.dirname(curPath))
+    const loadedModule = loadNodeModules(requirePath, posix.dirname(currentPath))
     if (loadedModule) return loadedModule
 
-    throw new Error(`Cannot find module ${reqPath}`)
+    throw new Error(`Cannot find module ${requirePath}`)
   }
 
   // create a top level require for the whole file system
-  const req = _require.bind(null, '')
+  const req = _require.bind(null, '/')
   req.extensions = extensions
   req.resolve = () => {}