+++ packages/examples/core/module-design/mountPlate.js
@@ -1,7 +1,0 @@
-const { cube } = require('@jscad/csg/api').primitives3d
-
-const mountPlate = (length) => {
-  return cube({ size: [length, 1, 1] })
-}
-
-module.exports = mountPlate